Add perf counters table
Add a `Layer` implementation that outputs the performance counter values for each of the span. Performance counters allow making initial decisions on what can be a performance bottleneck, e.g. memory access, branch prediction rate, context switches, etc. However `perf` and other tools are usually able to print the overall counter values while we are more interested to get the values for the span intervals.

`PrintPerfCountersLayer` prints for each span a table with all the counters it was created for. The set of the counters is configured via a vector of pairs (name, event) passed to the constructor. The reason why this functionality is not integrated into `PrintTreeLayer` is because in case when there is a dozen of counters the tree becomes too big and barely readable. So it is supposed that the user first uses the tree output to find out what are the spans of interest and finds the values of the counters corresponding to them in the table.